NH-76996 Otelcol: collecting Lambda function logs (code refactoring) (#21)
* Removed logs.go and logs_test.go and applied sharedcomponent to share the same http server with receiver.go
* Fixed code indexing problem by putting telemetryapi code to receiver/telemetryapireceiver/internal
* Enable WithMetrics with empty implementation so as to prepare for NH-70292
