+++ collector/receiver/telemetryapireceiver/README.md
@@ -1,10 +1,10 @@
 # Telemetry API Receiver
 
-| Status                   |                 |
-| ------------------------ |-----------------|
-| Stability                | [in development]|
-| Supported pipeline types | traces          |
-| Distributions            | [extension]     |
+| Status                   |                       |
+| ------------------------ |-----------------------|
+| Stability                | [in development]      |
+| Supported pipeline types | traces, metrics, logs |
+| Distributions            | [extension]           |
 
 This receiver generates telemetry in response to events from the [Telemetry API](https://docs.aws.amazon.com/lambda/latest/dg/telemetry-api.html). It does this by setting up an endpoint and registering itself with the Telemetry API on startup.
