+++ collector/receiver/telemetryapireceiver/go.mod
@@ -4,11 +4,8 @@ go 1.21.0
 
 toolchain go1.22.2
 
-replace github.com/open-telemetry/opentelemetry-lambda/collector => ../../
-
 require (
 	github.com/golang-collections/go-datastructures v0.0.0-20150211160725-59788d5eb259
-	github.com/open-telemetry/opentelemetry-lambda/collector v0.98.0
 	github.com/stretchr/testify v1.9.0
 	go.opentelemetry.io/collector/component v0.99.0
 	go.opentelemetry.io/collector/consumer v0.99.0
